Update ConvertIntToUnit() to use SI units
Adding in the `i` in units like
KiB: Kibibytes
GiB: Gigbibytes

Of course byte counts are in units of 1024, you would have to be an electrical engineer or a consumer to think something otherwise.  But I saw this when I was exploring the code again and thought I'd propose to patch it.